split out tests which fail on XML entities and DTD parsing
previously behind NO_XMLENTITIES and NO_XMLDOM flag, apparently fully unsupported by .NET Core/Standard
