+++ Testing/unittest/Parsing/Handlers/GraphHandlerTests.cs
@@ -35,62 +35,8 @@ using VDS.RDF.Writing.Formatting;
 
 namespace VDS.RDF.Parsing.Handlers
 {
-
-    public class GraphHandlerTests
+    public partial class GraphHandlerTests
     {
-#if !NO_SYNC_HTTP
-        [Fact]
-        public void ParsingGraphHandlerImplicitBaseUriPropogation()
-        {
-            try
-            {
-#if NET40
-                Options.UriLoaderCaching = false;
-#endif
-
-                Graph g = new Graph();
-                UriLoader.Load(g, new Uri("http://wiki.rkbexplorer.com/id/void"));
-                NTriplesFormatter formatter = new NTriplesFormatter();
-                foreach (Triple t in g.Triples)
-                {
-                    Console.WriteLine(t.ToString());
-                }
-            }
-            finally
-            {
-#if NET40
-                Options.UriLoaderCaching = true;
-#endif
-            }
-        }
-
-        [Fact]
-        public void ParsingGraphHandlerImplicitBaseUriPropogation2()
-        {
-            try
-            {
-#if NET40
-                Options.UriLoaderCaching = false;
-#endif
-
-                Graph g = new Graph();
-                g.LoadFromEmbeddedResource("VDS.RDF.Configuration.configuration.ttl");
-                UriLoader.Load(g, new Uri("http://wiki.rkbexplorer.com/id/void"));
-                NTriplesFormatter formatter = new NTriplesFormatter();
-                foreach (Triple t in g.Triples)
-                {
-                    Console.WriteLine(t.ToString());
-                }
-            }
-            finally
-            {
-#if NET40
-                Options.UriLoaderCaching = true;
-#endif
-            }
-        }
-#endif
-
         [Fact]
         public void ParsingGraphHandlerImplicitTurtle()
         {
@@ -156,14 +102,6 @@ namespace VDS.RDF.Parsing.Handlers
         {
             this.ParsingUsingGraphHandlerExplicitTest("graph_handler_tests_temp.n3", new Notation3Parser(), true);
         }
-
-#if !NO_XMLENTITIES
-        [Fact]
-        public void ParsingGraphHandlerExplicitRdfXml()
-        {
-            this.ParsingUsingGraphHandlerExplicitTest("graph_handler_tests_temp.rdf", new RdfXmlParser(), true);
-        }
-#endif
 
         [Fact]
         public void ParsingGraphHandlerExplicitRdfA()
