+++ Testing/unittest/Parsing/Handlers/ResultSetHandlerTests.NetFull.cs
@@ -1,0 +1,80 @@
+/*
+dotNetRDF is free and open source software licensed under the MIT License
+
+-----------------------------------------------------------------------------
+
+Copyright (c) 2009-2012 dotNetRDF Project (dotnetrdf-developer@lists.sf.net)
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is furnished
+to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using Xunit;
+using VDS.RDF.Parsing;
+using VDS.RDF.Parsing.Handlers;
+using VDS.RDF.Query;
+using VDS.RDF.Writing;
+using VDS.RDF.Writing.Formatting;
+
+namespace VDS.RDF.Parsing.Handlers
+{
+    public partial class ResultSetHandlerTests
+    {
+        [Fact]
+        public void ParsingResultSetHandlerImplicitSparqlRdfXml()
+        {
+            this.EnsureTestData("test.sparql.rdf", new SparqlRdfWriter(new RdfXmlWriter()));
+
+            SparqlRdfParser parser = new SparqlRdfParser(new RdfXmlParser());
+            SparqlResultSet results = new SparqlResultSet();
+            parser.Load(results, "test.sparql.rdf");
+
+            NTriplesFormatter formatter = new NTriplesFormatter();
+            foreach (SparqlResult r in results)
+            {
+                Console.WriteLine(r.ToString(formatter));
+            }
+
+            Assert.False(results.IsEmpty, "Result Set should not be empty");
+            Assert.Equal(SparqlResultsType.VariableBindings, results.ResultsType);
+        }
+
+        [Fact]
+        public void ParsingResultSetHandlerExplicitSparqlRdfXml()
+        {
+            this.EnsureTestData("test.sparql.rdf", new SparqlRdfWriter(new RdfXmlWriter()));
+
+            SparqlRdfParser parser = new SparqlRdfParser(new RdfXmlParser());
+            SparqlResultSet results = new SparqlResultSet();
+            parser.Load(new ResultSetHandler(results), "test.sparql.rdf");
+
+            NTriplesFormatter formatter = new NTriplesFormatter();
+            foreach (SparqlResult r in results)
+            {
+                Console.WriteLine(r.ToString(formatter));
+            }
+
+            Assert.False(results.IsEmpty, "Result Set should not be empty");
+            Assert.Equal(SparqlResultsType.VariableBindings, results.ResultsType);
+        }
+    }
+}
