+++ Testing/unittest/Parsing/Handlers/ResultSetHandlerTests.cs
@@ -37,8 +37,7 @@ using VDS.RDF.Writing.Formatting;
 
 namespace VDS.RDF.Parsing.Handlers
 {
-
-    public class ResultSetHandlerTests
+    public partial class ResultSetHandlerTests
     {
         private void EnsureTestData(String file)
         {
@@ -156,25 +155,6 @@ namespace VDS.RDF.Parsing.Handlers
         }
 
         [Fact]
-        public void ParsingResultSetHandlerImplicitSparqlRdfXml()
-        {
-            this.EnsureTestData("test.sparql.rdf", new SparqlRdfWriter(new RdfXmlWriter()));
-
-            SparqlRdfParser parser = new SparqlRdfParser(new RdfXmlParser());
-            SparqlResultSet results = new SparqlResultSet();
-            parser.Load(results, "test.sparql.rdf");
-
-            NTriplesFormatter formatter = new NTriplesFormatter();
-            foreach (SparqlResult r in results)
-            {
-                Console.WriteLine(r.ToString(formatter));
-            }
-
-            Assert.False(results.IsEmpty, "Result Set should not be empty");
-            Assert.Equal(SparqlResultsType.VariableBindings, results.ResultsType);
-        }
-
-        [Fact]
         public void ParsingResultSetHandlerImplicitSparqlRdfJson()
         {
             this.EnsureTestData("test.sparql.json", new SparqlRdfWriter(new RdfJsonWriter()));
@@ -277,25 +257,6 @@ namespace VDS.RDF.Parsing.Handlers
             SparqlRdfParser parser = new SparqlRdfParser(new Notation3Parser());
             SparqlResultSet results = new SparqlResultSet();
             parser.Load(new ResultSetHandler(results), "test.sparql.n3");
-
-            NTriplesFormatter formatter = new NTriplesFormatter();
-            foreach (SparqlResult r in results)
-            {
-                Console.WriteLine(r.ToString(formatter));
-            }
-
-            Assert.False(results.IsEmpty, "Result Set should not be empty");
-            Assert.Equal(SparqlResultsType.VariableBindings, results.ResultsType);
-        }
-
-        [Fact]
-        public void ParsingResultSetHandlerExplicitSparqlRdfXml()
-        {
-            this.EnsureTestData("test.sparql.rdf", new SparqlRdfWriter(new RdfXmlWriter()));
-
-            SparqlRdfParser parser = new SparqlRdfParser(new RdfXmlParser());
-            SparqlResultSet results = new SparqlResultSet();
-            parser.Load(new ResultSetHandler(results), "test.sparql.rdf");
 
             NTriplesFormatter formatter = new NTriplesFormatter();
             foreach (SparqlResult r in results)
