+++ Testing/unittest/Parsing/RdfXmlTests.NetFull.cs
@@ -1,0 +1,125 @@
+/*
+dotNetRDF is free and open source software licensed under the MIT License
+
+-----------------------------------------------------------------------------
+
+Copyright (c) 2009-2012 dotNetRDF Project (dotnetrdf-developer@lists.sf.net)
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is furnished
+to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR 
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, 
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+*/
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Xunit;
+using VDS.RDF.Parsing.Handlers;
+using VDS.RDF.Writing;
+using VDS.RDF.Writing.Formatting;
+
+namespace VDS.RDF.Parsing
+{
+    public partial class RdfXmlTests
+    {
+        [Fact]
+        public void ParsingRdfXmlEmptyStrings()
+        {
+            NTriplesFormatter formatter = new NTriplesFormatter();
+            RdfXmlParser domParser = new RdfXmlParser(RdfXmlParserMode.DOM);
+            Graph g = new Graph();
+            domParser.Load(g, "resources\\empty-string-rdfxml.rdf");
+
+            Console.WriteLine("DOM Parser parsed OK");
+
+            foreach (Triple t in g.Triples)
+            {
+                Console.WriteLine(t.ToString(formatter));
+            }
+            Console.WriteLine();
+
+            RdfXmlParser streamingParser = new RdfXmlParser(RdfXmlParserMode.Streaming);
+            Graph h = new Graph();
+            streamingParser.Load(h, "resources\\empty-string-rdfxml.rdf");
+
+            Console.WriteLine("Streaming Parser parsed OK");
+
+            foreach (Triple t in h.Triples)
+            {
+                Console.WriteLine(t.ToString(formatter));
+            }
+
+            Assert.Equal(g, h);
+        }
+
+        [Fact]
+        public void ParsingRdfXmlWithUrlEscapedNodes()
+        {
+            //Originally submitted by Rob Styles as part of CORE-251, modified somewhat during debugging process
+            NTriplesFormatter formatter = new NTriplesFormatter();
+            RdfXmlParser domParser = new RdfXmlParser(RdfXmlParserMode.DOM);
+            Graph g = new Graph();
+            domParser.Load(g, "resources\\urlencodes-in-rdfxml.rdf");
+
+            foreach (Triple t in g.Triples)
+            {
+                Console.WriteLine(t.ToString(formatter));
+            }
+
+            Uri encoded = new Uri("http://example.com/some%40encoded%2FUri");
+            Uri unencoded = new Uri("http://example.com/some@encoded/Uri");
+
+            Assert.False(EqualityHelper.AreUrisEqual(encoded, unencoded), "URIs should not be equivalent because %40 encodes a reserved character and per RFC 3986 decoding this can change the meaning of the URI");
+
+            IUriNode encodedNode = g.GetUriNode(encoded);
+            Assert.NotNull(encodedNode);
+            IUriNode unencodedNode = g.GetUriNode(unencoded);
+            Assert.NotNull(unencodedNode);
+
+            IUriNode pred = g.CreateUriNode(new Uri("http://example.org/schema/encoded"));
+            Assert.True(g.ContainsTriple(new Triple(encodedNode, pred, g.CreateLiteralNode("true"))), "The encoded node should have the property 'true' from the file");
+            Assert.True(g.ContainsTriple(new Triple(unencodedNode, pred, g.CreateLiteralNode("false"))), "The unencoded node should have the property 'false' from the file");
+        }
+
+        [Fact]
+        public void ParsingRdfXmlWithUrlEscapedNodes2()
+        {
+            //Originally submitted by Rob Styles as part of CORE-251, modified somewhat during debugging process
+            NTriplesFormatter formatter = new NTriplesFormatter();
+            RdfXmlParser domParser = new RdfXmlParser(RdfXmlParserMode.DOM);
+            Graph g = new Graph();
+            domParser.Load(g, "resources\\urlencodes-in-rdfxml.rdf");
+
+            foreach (Triple t in g.Triples)
+            {
+                Console.WriteLine(t.ToString(formatter));
+            }
+
+            Uri encoded = new Uri("http://example.com/some%20encoded%2FUri");
+            Uri unencoded = new Uri("http://example.com/some encoded/Uri");
+
+            IUriNode encodedNode = g.GetUriNode(encoded);
+            Assert.NotNull(encodedNode);
+            IUriNode unencodedNode = g.GetUriNode(unencoded);
+            Assert.NotNull(unencodedNode);
+
+            IUriNode pred = g.CreateUriNode(new Uri("http://example.org/schema/encoded"));
+            Assert.True(g.ContainsTriple(new Triple(encodedNode, pred, g.CreateLiteralNode("true"))), "The encoded node should have the property 'true' from the file");
+            Assert.True(g.ContainsTriple(new Triple(unencodedNode, pred, g.CreateLiteralNode("false"))), "The unencoded node should have the property 'false' from the file");
+
+        }
+    }
+}
