+++ Testing/unittest/Writing/CollectionCompressionTests.cs
@@ -37,287 +37,25 @@ using VDS.RDF.Writing.Formatting;
 
 namespace VDS.RDF.Writing
 {
-
-    public class CollectionCompressionTests
+    public partial class CollectionCompressionTests
         : CompressionTests
     {
-        [Fact]
-        public void WritingCollectionCompressionEmpty1()
+        [Fact(Skip = "Commented out before, now fails (?)")]
+        public void WritingCollectionCompressionSimple7()
         {
             Graph g = new Graph();
             g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
             INode n = g.CreateBlankNode();
             INode rdfType = g.CreateUriNode("rdf:type");
 
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(1, context.Collections.Count);
-            Assert.Equal(0, context.Collections.First().Value.Triples.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionEmpty2()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode rdfType = g.CreateUriNode("rdf:type");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), g.CreateUriNode("rdf:nil"));
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionSimple1()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateBlankNode();
-            INode rdfType = g.CreateUriNode("rdf:type");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(n, rdfType, g.CreateUriNode("ex:BlankNode"));
+            g.Assert(n, rdfType, g.CreateUriNode("ex:Obj"));
+            g.Assert(n, rdfType, g.CreateUriNode("ex:Test"));
 
             CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
             WriterHelper.FindCollections(context);
 
             Assert.Equal(1, context.Collections.Count);
             Assert.Equal(1, context.Collections.First().Value.Triples.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionSimple2()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateBlankNode();
-            INode rdfType = g.CreateUriNode("rdf:type");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred2"), n);
-            g.Assert(n, rdfType, g.CreateUriNode("ex:BlankNode"));
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionSimple3()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateBlankNode();
-            INode rdfType = g.CreateUriNode("rdf:type");
-            INode rdfFirst = g.CreateUriNode("rdf:first");
-            INode rdfRest = g.CreateUriNode("rdf:rest");
-            INode rdfNil = g.CreateUriNode("rdf:nil");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(n, rdfFirst, g.CreateLiteralNode("first"));
-            g.Assert(n, rdfRest, rdfNil);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(1, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionSimple4()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateBlankNode();
-            INode rdfType = g.CreateUriNode("rdf:type");
-            INode rdfFirst = g.CreateUriNode("rdf:first");
-            INode rdfRest = g.CreateUriNode("rdf:rest");
-            INode rdfNil = g.CreateUriNode("rdf:nil");
-
-            g.Assert(n, rdfFirst, g.CreateLiteralNode("first"));
-            g.Assert(n, rdfRest, rdfNil);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionSimple5()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateBlankNode();
-            INode rdfType = g.CreateUriNode("rdf:type");
-            INode rdfFirst = g.CreateUriNode("rdf:first");
-            INode rdfRest = g.CreateUriNode("rdf:rest");
-            INode rdfNil = g.CreateUriNode("rdf:nil");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred2"), n);
-            g.Assert(n, rdfFirst, g.CreateLiteralNode("first"));
-            g.Assert(n, rdfRest, rdfNil);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        //[Fact]
-        //public void WritingCollectionCompressionSimple6()
-        //{
-        //    Graph g = new Graph();
-        //    g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-        //    INode n = g.CreateBlankNode();
-        //    INode rdfType = g.CreateUriNode("rdf:type");
-
-        //    g.Assert(n, rdfType, g.CreateUriNode("ex:Obj"));
-
-        //    CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-        //    WriterHelper.FindCollections(context);
-
-        //    Assert.Equal(1, context.Collections.Count);
-        //    Assert.Equal(0, context.Collections.First().Value.Triples.Count, "Expected 1 Triple to be in the collection");
-
-        //    this.CheckCompressionRoundTrip(g);
-        //}
-
-        //[Fact]
-        //public void WritingCollectionCompressionSimple7()
-        //{
-        //    Graph g = new Graph();
-        //    g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-        //    INode n = g.CreateBlankNode();
-        //    INode rdfType = g.CreateUriNode("rdf:type");
-
-        //    g.Assert(n, rdfType, g.CreateUriNode("ex:Obj"));
-        //    g.Assert(n, rdfType, g.CreateUriNode("ex:Test"));
-
-        //    CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-        //    WriterHelper.FindCollections(context);
-
-        //    Assert.Equal(1, context.Collections.Count);
-        //    Assert.Equal(1, context.Collections.First().Value.Triples.Count, "Expected 1 Triple to be in the collection");
-
-        //    this.CheckCompressionRoundTrip(g);
-        //}
-
-#if !NO_SYNC_HTTP // No SparqlConnector
-        [Fact]
-        public void WritingCollectionCompressionComplex1()
-        {
-            SparqlConnector connector = new SparqlConnector(new VDS.RDF.Query.SparqlRemoteEndpoint(new Uri("http://dbpedia.org/sparql")));
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            g.NamespaceMap.AddNamespace("dnr", new Uri(ConfigurationLoader.ConfigurationNamespace));
-            INode n = g.CreateBlankNode();
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("dnr:genericManager"), n);
-            ConfigurationSerializationContext sContext = new ConfigurationSerializationContext(g);
-            sContext.NextSubject = n;
-            connector.SerializeConfiguration(sContext);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(2, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-#endif
-
-        [Fact]
-        public void WritingCollectionCompressionComplex2()
-        {
-            Graph g = new Graph();
-            g.LoadFromFile("resources\\complex-collections.nt");
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            NTriplesFormatter formatter = new NTriplesFormatter();
-            foreach (KeyValuePair<INode, OutputRdfCollection> kvp in context.Collections)
-            {
-                Console.WriteLine("Collection Root - " + kvp.Key.ToString(formatter));
-                Console.WriteLine("Collection Triples (" + kvp.Value.Triples.Count + ")");
-                foreach (Triple t in kvp.Value.Triples)
-                {
-                    Console.WriteLine(t.ToString(formatter));
-                }
-                Console.WriteLine();
-            }
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionNamedListNodes()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateUriNode("ex:list");
-            INode rdfType = g.CreateUriNode("rdf:type");
-            INode rdfFirst = g.CreateUriNode("rdf:first");
-            INode rdfRest = g.CreateUriNode("rdf:rest");
-            INode rdfNil = g.CreateUriNode("rdf:nil");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(n, rdfFirst, g.CreateLiteralNode("first"));
-            g.Assert(n, rdfRest, rdfNil);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
-
-            this.CheckCompressionRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingCollectionCompressionNamedListNodes2()
-        {
-            Graph g = new Graph();
-            g.NamespaceMap.AddNamespace("ex", new Uri("http://example.org/"));
-            INode n = g.CreateUriNode("ex:listRoot");
-            INode m = g.CreateUriNode("ex:listItem");
-            INode rdfType = g.CreateUriNode("rdf:type");
-            INode rdfFirst = g.CreateUriNode("rdf:first");
-            INode rdfRest = g.CreateUriNode("rdf:rest");
-            INode rdfNil = g.CreateUriNode("rdf:nil");
-
-            g.Assert(g.CreateUriNode("ex:subj"), g.CreateUriNode("ex:pred"), n);
-            g.Assert(n, rdfFirst, g.CreateLiteralNode("first"));
-            g.Assert(n, rdfRest, m);
-            g.Assert(m, rdfFirst, g.CreateLiteralNode("second"));
-            g.Assert(m, rdfRest, rdfNil);
-
-            CompressingTurtleWriterContext context = new CompressingTurtleWriterContext(g, Console.Out);
-            WriterHelper.FindCollections(context);
-
-            Assert.Equal(0, context.Collections.Count);
 
             this.CheckCompressionRoundTrip(g);
         }
