+++ Testing/unittest/Writing/RdfXmlWriterTests.cs
@@ -37,9 +37,7 @@ using VDS.RDF.Writing.Formatting;
 
 namespace VDS.RDF.Writing
 {
-#if !NO_XMLENTITIES // RDF/XML output generates entity declarations that the parser cannot process
-
-    public class RdfXmlWriterTests
+    public partial class RdfXmlWriterTests
     {
         private List<IRdfWriter> _writers = new List<IRdfWriter>()
         {
@@ -280,37 +278,6 @@ namespace VDS.RDF.Writing
         }
 
         [Fact]
-        public void WritingRdfXmlSimpleCollection()
-        {
-            String fragment = "@prefix : <http://example.org/>. :subj :pred ( 1 2 3 ).";
-
-            Graph g = new Graph();
-            g.LoadFromString(fragment);
-
-            this.CheckRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingRdfXmlComplex()
-        {
-            Graph g = new Graph();
-            TurtleParser parser = new TurtleParser();
-            parser.Load(new PagingHandler(new GraphHandler(g), 1000), "resources\\chado-in-owl.ttl");
-
-            this.CheckRoundTrip(g);
-        }
-
-        [Fact]
-        public void WritingRdfXmlWithDtds()
-        {
-            String fragment = "@prefix xsd: <" + NamespaceMapper.XMLSCHEMA + ">. @prefix : <http://example.org/>. :subj a :obj ; :has \"string\"^^xsd:string ; :has 23 .";
-            Graph g = new Graph();
-            g.LoadFromString(fragment);
-
-            this.CheckRoundTrip(g);
-        }
-
-        [Fact]
         public void WritingRdfXmlInvalidPredicates1()
         {
             String fragment = "@prefix ex: <http://example.org/>. ex:subj ex:123 ex:object .";
@@ -356,5 +323,4 @@ namespace VDS.RDF.Writing
             Assert.True(outData.Contains("rdf:resource=\"&ex;2o\""));
         }
     }
-#endif
 }
